# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "WaterIn server",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "App for monitoring the amount of water consumed"
  },
  "servers": [
    {
      "url": "https://waterin-server.onrender.com"
    }
  ],
...
  "paths": {
...
    "/water/day/{time}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get water notes for day",
        "operationId": "getWaterPerDay",
        "description": "Get water notes per day",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "time",
            "required": true,
            "schema": {
              "type": "number",
              "example": 1722772947000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully find notes on day",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully find water notes!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "notes": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "userId",
                              "volume",
                              "drinkTime"
                            ],
                            "properties": {
                              "userId": {
                                "type": "string",
                                "example": "1234567890abcdef"
                              },
                              "volume": {
                                "type": "number",
                                "example": 500
                              },
                              "drinkTime": {
                                "type": "string",
                                "example": "12:45"
                              },
                              "createdAt": {
                                "type": "number",
                                "example": 1657123456789
                              },
                              "updatedAt": {
                                "type": "number",
                                "example": 1657123456789
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/water/month/{time}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get water notes for month",
        "operationId": "getWaterPerMonth",
        "description": "Get water notes per month divided on days",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "time",
            "required": true,
            "schema": {
              "type": "number",
              "example": 1722772947000
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "water": {
        "type": "object",
        "required": [
          "userId",
          "volume",
          "drinkTime"
        ],
        "properties": {
          "userId": {
            "description": "ID of the user who recorded the water intake",
            "type": "string",
            "example": "64a8c29f8c1d8d1a4e8b4567"
          },
          "volume": {
            "description": "Volume of water consumed in milliliters",
            "type": "number",
            "example": 500
          },
          "drinkTime": {
            "description": "Time when the water was consumed",
            "type": "string",
            "format": "time",
            "example": "15:30"
          },
          "createdAt": {
            "description": "Time in Unix format when the note was created",
            "type": "integer",
            "format": "int64",
            "example": 1699209600000
          },
          "updatedAt": {
            "description": "Time in Unix format when the note was updated",
            "type": "integer",
            "format": "int64",
            "example": 1699889600000
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Successfully found water notes",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successful find water notes by date"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "date",
                    "waterNotes"
                  ],
                  "properties": {
                    "date": {
                      "type": "string",
                      "example": "2024-08-04",
                      "description": "Date of month where water notes were searched"
                    },
                    "waterNotes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "$ref": "#/components/schemas/water"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "404": {
        "description": "Not Found Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Entity not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}